Remove top padding from container
The padding was messing with the even spacing of
the new heading style. So the top padding has been
moved to its own class.

This means only the footer and profile pages
needed that class to add the padding back
in.

If I made the top padding the container default,
I would have had to go through every page to
remove the top-padding from each page's main
container. Boring!